refactor(plugin-space): move CreateObjectPanel and gate registry button on type selector
Move CreateObjectPanel into its own directory and show the create-object
dialog registry button only while the type selector is visible.
