+++ packages/plugins/plugin-space/src/components/index.ts
@@ -5,7 +5,7 @@
 import { type ComponentType, lazy } from 'react';
 
 export * from './AwaitingObject';
-export * from './CreateDialog';
+export * from './CreateObjectPanel';
 export * from './ForeignKeys';
 export * from './SyncStatus';
